% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initbeta.R
\name{init.beta}
\alias{init.beta}
\title{FUNCTION_TITLE}
\usage{
init.beta(dat, method = c("glmnet", "lme4", "glmmPQL"), ...)
}
\arguments{
\item{dat}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #init.beta(sim.data[[1]])
 }
}
}
\seealso{
\code{\link[glmnet]{cv.glmnet}}
}
