% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_section.R
\name{golden_section}
\alias{golden_section}
\title{FUNCTION_TITLE}
\usage{
golden_section(data, init.beta, pen.effect = "FE.L1", opt.lb = 0,
  opt.ub = 1, opt.maxiter = 100, opt.tol = 0.1, opt.tau = (sqrt(5) -
  1)/2)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{init.beta}{PARAM_DESCRIPTION}

\item{pen.effect}{PARAM_DESCRIPTION, Default: 'FE.L1'}

\item{opt.lb}{PARAM_DESCRIPTION, Default: 0}

\item{opt.ub}{PARAM_DESCRIPTION, Default: 1}

\item{opt.maxiter}{PARAM_DESCRIPTION, Default: 100}

\item{opt.tol}{PARAM_DESCRIPTION, Default: 0.1}

\item{opt.tau}{numeric, golden proportion coefficient (~0.618) Default: (sqrt(5) - 1)/2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 dat <- initialize_example(n.i = 5,n = 30,q=4,seed=1)
 init <- init.beta(dat,method='glmnet')
 golden_section(dat,init,pen.effect = 'FE.L1')
 }
}
}
