% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmen.R
\name{lmmen}
\alias{lmmen}
\title{linear mixed model elastic net}
\usage{
lmmen(data, init.beta, frac, eps = 10^(-4), verbose = FALSE)
}
\arguments{
\item{data}{matrix, data}

\item{init.beta}{numeric, initial values for fixed effects coefficients}

\item{frac}{PARAM_DESCRIPTION}

\item{eps}{numeric, tolerance level to pass to solve.QP, Default: 10^(-4)}

\item{verbose}{boolean, show output during optimization Default: FALSE}
}
\value{
list
}
\description{
Solve a linear mixed model elastic net optimization problem.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 EXAMPLE1
 }
}
}
